[dawn][native] Use strong types for Command structs
Specifically, modify TextureCopy, CopyBufferToTextureCmd,
CopyTextureToBufferCmd, and CopyTextureToTextureCmd structs to use
strong types (TexelOrigin3D and TexelExtent3D), and update code that
uses them.

Also opportunistically update other code to use strong types.

Bug: 424536624
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/276714
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Auto-Submit: Antonio Maiorano <amaiorano@google.com>

+++ src/dawn/native/BlitTextureToBuffer.cpp
@@ -1238,9 +1238,9 @@ MaybeError BlitTextureToBuffer(DeviceBase* device,
         uint32_t* params =
             static_cast<uint32_t*>(uniformBuffer->GetMappedRange(0, bufferDesc.size));
         // srcOrigin: vec3u
-        params[0] = src.origin.x;
-        params[1] = src.origin.y;
-        params[2] = src.origin.z;
+        params[0] = static_cast<uint32_t>(src.origin.x);
+        params[1] = static_cast<uint32_t>(src.origin.y);
+        params[2] = static_cast<uint32_t>(src.origin.z);
 
         // packTexelCount: number of texel values (1, 2, or 4) one thread packs into the dst
         // buffer
